+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- feat: Collector metadata tag auto-discovery (local services, e.g. mysql) [#893]
+
+[Unreleased]: https://github.com/SumoLogic/sumologic-otel-collector/compare/v0.69.0-sumo-0...main
+[#893]: https://github.com/SumoLogic/sumologic-otel-collector/pull/893
+
 ## [v0.69.0-sumo-0]
 
 ### Added